requirements.txt
# AI Recruiting Demo - Python Dependencies
# Generated: 2025-11-13
# Python 3.11+

# Core Framework
flask==3.0.0
werkzeug==3.0.1
pyyaml==6.0.3

# PDF Processing
pdfplumber==0.11.0

# AI/LLM Integration
openai>=2.0.0
tiktoken==0.6.0

# Configuration & Environment
python-dotenv==1.0.0

# Optional: Email Support (Epic 8)
# flask-mail==0.9.1

